Refactor current declaration to be indiviudal
