+++ app/models/declaration/IndividualDeclaration.scala
@@ -19,11 +19,10 @@ package models.declaration
 import models.NameType
 import play.api.libs.json.{Json, OFormat}
 
-final case class Declaration(name: NameType, email: Option[String]) {
-
+final case class IndividualDeclaration(name: NameType, email: Option[String]) {
   def toJson = Json.toJson(this)
 }
 
-object Declaration {
-  implicit lazy val formats: OFormat[Declaration] = Json.format[Declaration]
+object IndividualDeclaration {
+  implicit lazy val formats: OFormat[IndividualDeclaration] = Json.format[IndividualDeclaration]
 }
