+++ app/pages/DeclarationPage.scala
@@ -16,10 +16,10 @@
 
 package pages
 
-import models.declaration.Declaration
+import models.declaration.IndividualDeclaration
 import play.api.libs.json.JsPath
 
-case object DeclarationPage extends QuestionPage[Declaration] {
+case object DeclarationPage extends QuestionPage[IndividualDeclaration] {
 
   override def path: JsPath = JsPath \ toString
