+++ conf/app.routes
@@ -40,6 +40,7 @@ GET        /last-declared                               controllers.print.LastDe
 GET        /declare-what-changed                        controllers.WhatIsNextController.onPageLoad()
 POST       /declare-what-changed                        controllers.WhatIsNextController.onSubmit()
 
-GET        /declare                                     controllers.declaration.DeclarationController.onPageLoad()
-POST       /declare                                     controllers.declaration.DeclarationController.onSubmit()
+GET        /declaration                                 controllers.declaration.IndividualDeclarationController.onPageLoad()
+POST       /declaration                                 controllers.declaration.IndividualDeclarationController.onSubmit()
+
 GET        /there-has-been-a-problem-declaring          controllers.declaration.ProblemDeclaringController.onPageLoad()
